pending: fix move queue

+++ Math/Position.cs
@@ -2,12 +2,12 @@ using System;
 
 namespace Formula.Math;
 
-public struct Vector2D(int x, int y)
+public struct Vector2D(double x, double y)
 {
-    public int X = x;
-    public int Y = y;
+    public double X = x;
+    public double Y = y;
 
-    public static Vector2D operator *(Vector2D v, int scalar)
+    public static Vector2D operator *(Vector2D v, double scalar)
         => new Vector2D(v.X * scalar, v.Y * scalar);
     
     public static bool operator ==(Vector2D a, Vector2D b)
